refactor(RepositoryTest) 테스트 내용이 Repository에 가까움에 따른 클래스명 변경
