+++ src/main/java/nextstep/subway/linebridge/ui/LineBridgeController.java
@@ -2,7 +2,7 @@ package nextstep.subway.linebridge.ui;
 
 import javax.websocket.server.PathParam;
 import nextstep.subway.linebridge.dto.LineBridgeDto;
-import nextstep.subway.linebridge.application.LineBridgeService;
+import nextstep.subway.linebridge.repotitory.LineBridgeService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
